# openapi.yaml
openapi: 3.0.3
info:
  title: WinOJ API
  description: |
    Windows 在线评测系统 API。支持多语言代码评测、安全沙箱执行、自定义计分脚本、比赛管理、文章系统、Rating 机制、AI 代码安全审查等功能。

    ## 认证
    大部分接口需携带 JWT Token：`Authorization: Bearer <access_token>`
    Access Token 有效期 15 分钟，过期后通过 `POST /auth/refresh` 刷新。

    ## 错误响应
    ```json
    { "code": 1|2|3|4|5|6, "reason": "ERR_xxx", "message": "..." }
    ```
    错误码：1=参数错误, 2=状态错误, 3=未找到, 4=限流, 5=未认证, 6=权限不足

    ## 角色层级
    user < teacher < admin < su
  version: 1.6.0
  contact:
    name: WinOJ
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 本地开发
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Problem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        input_desc:
          type: string
        output_desc:
          type: string
        hint:
          type: string
        time_limit:
          type: integer
          description: 毫秒
        memory_limit:
          type: integer
          description: MB
        problem_type:
          type: string
          enum: [traditional, interactive, communication, submit_answer]
        compare_mode:
          type: string
          enum: [text_strict, text_relaxed, real_number, spj]
        is_public:
          type: integer
          enum: [0, 1]
        provider:
          type: string
        created_at:
          type: string
    Submission:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        problem_id:
          type: integer
        language:
          type: string
        source_code:
          type: string
        status:
          type: string
          enum: [pending, running, compiling, judging, accepted, wrong_answer, time_limit_exceeded, memory_limit_exceeded, runtime_error, compile_error, system_error, pending_rejudge]
        score:
          type: number
        time_used:
          type: integer
          description: 毫秒
        memory_used:
          type: integer
          description: KB
        created_at:
          type: string
    SubmissionDetail:
      type: object
      properties:
        id:
          type: integer
        submission_id:
          type: integer
        test_case_id:
          type: integer
        group_id:
          type: integer
        subtask_id:
          type: string
        status:
          type: string
        score:
          type: number
        time_used:
          type: integer
        memory_used:
          type: integer
        stdout:
          type: string
        stderr:
          type: string
        exit_code:
          type: integer
...
    Tag:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        created_at:
          type: string
...
    Language:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        display_name:
          type: string
        compile_cmd:
          type: string
        run_cmd:
          type: string
        extension:
          type: string
        is_enabled:
          type: integer
          enum: [0, 1]
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
    SearchParam:
      name: search
      in: query
      schema:
        type: string
...
paths:
...
  /problems:
    get:
      tags: [题目]
      summary: 题目列表
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: tag
          in: query
          schema:
            type: string
          description: 按标签筛选
      responses:
        '200':
          description: 题目列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  problems:
                    type: array
                    items:
                      $ref: '#/components/schemas/Problem'
    post:
      tags: [题目]
      summary: 创建题目
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                input_desc:
                  type: string
                output_desc:
                  type: string
                hint:
                  type: string
                time_limit:
                  type: integer
                  default: 1000
                memory_limit:
                  type: integer
                  default: 256
                problem_type:
                  type: string
                  enum: [traditional, interactive, communication, submit_answer]
                  default: traditional
                compare_mode:
                  type: string
                  enum: [text_strict, text_relaxed, real_number, spj]
                  default: text_strict
                real_number_tolerance:
                  type: object
                  properties:
                    absolute:
                      type: number
                      default: 0.001
                    relative:
                      type: number
                      default: 0.001
                spj_code:
                  type: string
                allowed_languages:
                  type: array
                  items:
                    type: string
                is_public:
                  type: boolean
                  default: true
                provider:
                  type: string
      responses:
        '201':
          description: 创建成功
...
  /tags:
    get:
      tags: [标签]
      summary: 标签列表
      responses:
        '200':
          description: 标签列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
    post:
      tags: [标签]
      summary: 创建标签
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                color:
                  type: string
                  default: "#6366f1"
      responses:
        '201':
          description: 创建成功
...
  /submissions:
    get:
      tags: [提交]
      summary: 提交列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: user_id
          in: query
          schema:
            type: integer
        - name: problem_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, compiling, judging, accepted, wrong_answer, time_limit_exceeded, memory_limit_exceeded, runtime_error, compile_error, system_error]
      responses:
        '200':
          description: 提交列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                  submissions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Submission'
    post:
      tags: [提交]
      summary: 提交代码
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [problem_id, language]
              properties:
                problem_id:
                  type: integer
                language:
                  type: string
                source_code:
                  type: string
                  maxLength: 131072
                answer_data:
                  type: string
      responses:
        '201':
          description: 提交成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  submission_id:
                    type: integer
                  message:
                    type: string
        '403':
          description: 安全审查失败，用户被封禁
  /submissions/{id}:
    get:
      tags: [提交]
      summary: 提交详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 提交详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Submission'
                  - type: object
                    properties:
                      details:
                        type: array
                        items:
                          $ref: '#/components/schemas/SubmissionDetail'
    delete:
      tags: [提交]
      summary: 删除提交
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
...
  /users:
    get:
      tags: [用户]
      summary: 用户列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - name: role
          in: query
          schema:
            type: string
            enum: [user, teacher, admin, su]
      responses:
        '200':
          description: 用户列表
    post:
      tags: [用户]
      summary: 创建用户
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                nickname:
                  type: string
                role:
                  type: string
                  enum: [user, teacher, admin, su]
                  default: user
      responses:
        '201':
          description: 创建成功
...
  /languages:
    get:
      tags: [语言]
      summary: 语言列表
      responses:
        '200':
          description: 语言列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Language'
    post:
      tags: [语言]
      summary: 添加语言
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, display_name, run_cmd, extension]
              properties:
                name:
                  type: string
                display_name:
                  type: string
                compile_cmd:
                  type: string
                run_cmd:
                  type: string
                extension:
                  type: string
      responses:
        '201':
          description: 添加成功
...
  /contests:
    get:
      tags: [比赛]
      summary: 比赛列表
      responses:
        '200':
          description: 比赛列表
    post:
      tags: [比赛]
      summary: 创建比赛
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, start_time, end_time]
              properties:
                title:
                  type: string
                description:
                  type: string
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                is_virtual:
                  type: integer
      responses:
        '201':
          description: 创建成功
...
  /articles:
    get:
      tags: [文章]
      summary: 文章列表
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 文章列表
    post:
      tags: [文章]
      summary: 发布文章
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                content:
                  type: string
                  description: Markdown 格式
                provider:
                  type: string
                is_published:
                  type: integer
                  enum: [0, 1]
                  default: 0
      responses:
        '201':
          description: 发布成功
...
  /uploads:
    get:
      tags: [文件]
      summary: 文件列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: 文件列表
    post:
      tags: [文件]
      summary: 上传文件
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: "最大 10MB"
      responses:
        '201':
          description: 上传成功
...
  /discussions:
    post:
      tags: [讨论]
      summary: 创建讨论
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content]
              properties:
                problem_id:
                  type: integer
                  description: 题目讨论，与 contest_id 二选一
                contest_id:
                  type: integer
                  description: 比赛讨论，与 problem_id 二选一
                title:
                  type: string
                content:
                  type: string
                  description: Markdown 格式
      responses:
        '201':
          description: 创建成功
...
  /categories:
    get:
      tags: [分类]
      summary: 分类列表
      responses:
        '200':
          description: 分类列表
    post:
      tags: [分类]
      summary: 创建分类
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: 创建成功
...
  /announcements:
    get:
      tags: [公告]
      summary: 公告列表
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            default: global
      responses:
        '200':
          description: 公告列表
    post:
      tags: [公告]
      summary: 创建公告
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                content:
                  type: string
                type:
                  type: string
                  default: global
                pinned:
                  type: boolean
                  default: false
      responses:
        '201':
          description: 创建成功
...
  /problem-sets:
    get:
      tags: [题单]
      summary: 题单列表
      responses:
        '200':
          description: 公开题单列表
    post:
      tags: [题单]
      summary: 创建题单
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                is_public:
                  type: integer
                  enum: [0, 1]
                  default: 1
                problemIds:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: 创建成功
...
  /jobs/{id}:
    get:
      tags: [系统]
      summary: 任务详情重定向
      description: 302 重定向到 /api/v1/submissions/{id}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '302':
          description: 重定向